+++ packages/koop-core/lib/provider-registration/create-model.js
@@ -82,9 +82,6 @@ module.exports = function createModel ({ ProviderModel, koop, namespace }, optio
       if (this.getStream) {
         await this.before(req)
         const providerStream = await this.getStream(req)
-        providerStream.on('end', () => {
-          this.after(req)
-        })
         return providerStream
       } else {
         throw new Error('getStream() function is not implemented in the provider.')